Update disallowed-resources.md
This page doesn't provide an example to clearly explain the effect of "Not allowed resource types" on a specific resource. Ironically, it's a question which is commonly asked in Azure Administrator certification exam. 

The effect of this parameter on VNet is mentioned in many 3rd party websites but hasn't been mentioned any where in the doc. This should be clearly stated in the doc as an official reference. Here is my suggestion to clarify it:

"If this parameter is set to virtualNetworks and scoped to a resource group, you can't create virtual networks within the resource group, you can't move virtual networks into the resource group, you can't create address space inside virtual network however you can create subnet."