Update app-service-plan-manage.md
Effect of moving a web app to a different subscription, hasn't been mentioned any where in the doc. I've added following for clarification.

'When you move an app to a different subscription, location of the app remains the same but its policy will be changed. For example, if your app is in Subscription1 located in Central US and has Policy1. And Subscription2 is in the UK South and has Policy2. If you move the app to Subscription2, the location of the app remains the same (Central US) however it will be under the new policy which is policy2.'