Update app-service-move-limitations.md
The following fact hasn't been mentioned any where in the doc so I've clarified it:

'When you move a Web App to a different subscription, location of the Web App remains the same but its policy will be changed. For example, if your Web App is in Subscription1 located in Central US and has Policy1. And Subscription2 is in the UK South and has Policy2. If you move the Web App to Subscription2, the location of the Web App remains the same (Central US) however it will be under the new policy which is policy2.'