Enhance clarity in workaround for unsupported connection string types
This commit expands the workaround note for configuring connection strings in Azure App Service to increase visibility and understanding of the necessary workaround for connection string types that are not handled by `EnvironmentVariablesConfigurationProvider`. Specifically, it clarifies that the workaround to use the "Custom" type applies not only to PostgreSQL but also to other services like Notification Hub, Service Bus, Event Hub, Document Db, and Redis Cache.