Update app-insights-overview.md
The following hasn't been explained any where in the doc, so I've clarified it through a FAQ:

"What's the difference between Application Insights Agent and SDKs?

Application Insights Agent is the easiest to enable; you only need to install the agent, there is no need for code changes or advanced configurations. Alternatively, you can manually instrument your code using Application Insights SDK. This means that you have to manage the updates to the latest version of the packages by yourself. Use SDK if you need to make custom dependency calls or API calls that are not captured by default with the agent."