Update managed-disks-overview.md
The following fact hasn't been mentioned anywhere in the doc so I've added it for clarity:

"On Windows VMs, drive C: is a persistent storage. This means that even if you redeploy the virtual machine, the data stored on this drive wouldn't be deleted. By default, new folders and user accounts are stored on drive C."