Update cache-remove-tls-10-11.md
There appears to be a discrepancy w.r.t. making TLS changes mentioned in the phase section versus what we have in the timelines.

Based on the sample test, it seems to be following the timeline and hence requesting to update this section as well. Please verify this once and review it further.