Update redundancy-migration.md
In many 3rd party websites and even Microsoft's own forum, there are many references to "live migration" (see examples below). I guess this is an old name but I see no mention of it in the doc, which is confusing. I've clarified it to prevent confusion:

"Conversion was formerly known as "live migration""

Few examples:
https://learn.microsoft.com/en-us/answers/questions/1215467/clarifications-required-on-storage-redundancy

https://learn.microsoft.com/en-us/answers/questions/311529/azure-storage-account-switch-between-types-of-repl