Update vpn-gateway-peering-gateway-transit.md
There should be an example of what counts as change in network topology. Actually, it's one of the questions asked in Azure Administrator certification exam, wherein they specifically mention that peering has been added and ask for troubleshooting. The answer, is to download and reinstall the package. 

If you make a change to the topology of your network (such as adding peering) and have Windows VPN clients, the VPN client package for Windows clients must be downloaded and installed again in order for the changes to be applied to the client.