Update role-definitions.md
Following should be added so readers are informed. Currently, this information is buried in documentation and requires pre-existing knowledge.

"Fully qualified unique ID of the role. Even if the role is renamed, the role ID does not change. It's a best practice to use the role ID in your scripts. "