Update role-assignments.md
Following should be added so readers are informed. Currently, this information requires pre-existing knowledge.

"The unique ID of the role assignment, which includes the name. Even if the role assignment is renamed, the role assignment ID does not change. It's a best practice to use the role assignment ID in your scripts."