Update custom-domain-external.md
Currently the information about the DNS records, their characteristics,  and how to use them with web app, is scattered around the doc and requires digging various pages to understand it. I personally had a hard time understanding it so I suggest adding the following note for better clarity:

"You can map a custom domain to your Azure Static Web App using either `CNAME` or `A` record. `CNAME` record maps a domain name to another domain (or subdomain) whereas `A` record maps a domain name to an IP address. If the IP address changes, a `CNAME` entry is still valid, unlike `A` record. Let’s say you have WebApp1 and you would like users to access it from https://www.contoso.com. You can do it in two possible ways: you can create a `CNAME` record, also known as an alias, and map it to WebApp1.azurestaticapps.net. Alternatively, you can create an `A` record and map it to the IP address of WebApp1."