Update dns-custom-domain.md
In this article, both CNAME and A records are mentioned. Following  should be added so readers have a better understanding of difference between the two records, so that they can choose either one more wisely.


"`CNAME` record maps a domain name to another domain (or subdomain) whereas `A` record maps a domain name to an IP address. If the IP address changes, a `CNAME` entry is still valid, unlike `A` record."