Update spot-vms.md
I'm afraid the current sentence is vague and confusing. I could understand it only because I read other sources. I would recommend to replace it with the following:

"Spot VMs can be stopped if Azure needs capacity for other pay-as-you-go workloads or when the price of the spot instance exceeds the maximum price that you have set. "