Update app-service-move-limitations.md
I'm afraid  this page is a bit confusing: in the title, it tells about moving to a resource group and subscription but down in the page, it has section only for, and mainly / only describes moving  to a subscription (not resource group).

AFAIK, the way it describes restrictions about moving TLS, isn't technically correct. Also, that applies to both subscription and resource group. So, I've modified and re-arranged that paragraph:


"You can move App Service resources to a new resource group or subscription but you need to delete and upload its TLS/SSL certificates to the new new resource group or subscription. However, you can't move a free App Service managed certificate. For that scenario, see [Move with free managed certificates](#move-with-free-managed-certificates).
"