Update quick-create-portal.md
The following hasn't been mentioned any where in the doc. Given the nature of this page, it should be mentioned here for clarity:

"To move a virtual machine from one virtual network to another, you need to delete and recreate the virtual machine in the new virtual network. However, you can keep its disks. "