Update vpn-gateway-about-vpngateways.md
Ironically, this page has no mention of VPN Gateway capabilities for VNet peering. It can give impression that VPN Gateway doesn't support it. I've added the following use case to disambiguate:

"Use VPN Gateway as a hub to connect with spokes such as peered networks, site-to-site connections, point-to-site connections and VNet-to-VNet connections."