[GH Issue] Update redhat-create-upload-vhd.md
There are specific V5 series of images in Azure for which Accelerated Networking is required. When one were to migrate to those series (mentioned below) of VMs, and if this udev rule is missing, then NetworkManager assigns the same ip address to Virtual Function interface as well. (direct quote from GH issue)