Update tutorial-load-balancer-ip-backend-portal.md
"whether a VM with public IP can be attached to a load balancer" is one of the points covered in Azure certification exams. Surprisingly, I couldn't find this fact anywhere in the doc (or it requires interpretation from some of the existing QuickStarts) so I had to test it in the portal by myself. I have added the following so documentation is clear about it:

"VMs with and without an associated public IP address can be attached to backend pool of a public load balancer."