Add additional token calculation information to the Limit Azure OpenAI API token usage API Management policy
How does the new APIM policy “Limit Azure OpenAI API token usage” `estimate-prompt-tokens` feature work? How does it estimate the tokens? Only of the prompt or also of the response? What changes, when I activate/deactivate it? The documentation was way too minimal on this.

This change adds some additional information about the impact of setting `estimate-prompt-tokens` to false.

This is important, because it might cause the probably unexpected behavior of sending requests to the LLM backend, which should have been filtered out by the policy.

Please check for correctness by the Product Team.