Update storage-account-key-note-include.md
Correcting the permissions that storage account keys give you as tested with Storage Explorer v1.34. Accessing a Storage account with the key gives you full access to the data, but not to the configuration or other keys. 
This include is causing confusion and contradicts other parts of the parent Learn doc: https://learn.microsoft.com/en-us/azure/storage/common/storage-account-keys-manage?tabs=azure-portal#protect-your-access-keys.