Update restore-sql-pool-from-deleted-workspace.md
I updated the below as the desired date time selection should not be an approximate it needs to be explicit, or it will fail. In addition, I updated the variable $DroppedDateTime to be the initialized variable to get the date time selection. The prior request was using the incorrect variable name.                                                                    

# Define the approximate point in time the workspace was dropped as DroppedDateTime "yyyy-MM-ddThh:mm:ssZ" (ex. 2022-01-01T16:15:00Z)
$PointInTime="<DroppedDateTime>" 
$DroppedDateTime = Get-Date -Date $PointInTime