Update python.md  to remove setting the keys and region every time we ran the code
I don't think its necessary for the setting of both environment variables to be called every time we call the function.

If the file will always be called from the command line, then no issues. But if we run the file in a text editor or an IDE, setting both environment variables once is enough, that way if we call the function, it will not again set those variables which are static.