Update restore-sql-pool-from-deleted-workspace.md
Correct the script to provide the correct code cmdlet that I used with the client which was Get-AzSynapseDroppedSqlPool. Added notes regarding that the Get-AzSynapseDroppedSqlPool needs to have the workspace available. If not, they can easily recreate the same workspace with the same values and the command will pick up the dropped database to be safely restored.