Update deployment.md
When I have tested this scenario & used the subnet range "10.1.0.0/16" mentioned in the document, the deployment is getting failed as the IP address is overlapping. The error I am getting is that "Subnet 'cloudshellsubnet' is not valid because its IP address range overlaps with that of an existing subnet in virtual network 'vnet-cloudshell-eastus'. (Code: NetcfgSubnetRangesOverlap)"