Update overview-diagnostics.md
First, this is a bit confusing. There is no search for symptoms, that's so vague. Next, that's the the next thing. Next adds negative value given it's not order dependent. Finally, finally should never be used.
Why give a poor summary of this, just send them to the right place. Finally, remember less is more per the MSFT style guide.