Update cognitive-services-support-options.md
We often get feedbacks that new customers cannot find "OpenAI" when they click "Create a support request" in "Help + support" because Azure Portal isn't fixed yet for a long time. So, this page should add the steps to raise a support ticket of Azure OpenAI Service. However, the page is redirected to Azure AI services' page. So, I wrote it as Azure AI services.