Removed or replaced work "you'll" as in description
Create Virtual Machine section:
Line 160 in Meta Data:
Replaced or removed the work "you'll"
Create a virtual machine in this section. During creation and add it to the backend pool of the internal load balancer. After the virtual machine is created, you can add the virtual machine to the backend pool of the public load balancer.


Add VM to backend pool of public load balancer
Line 215 in Meta Data
Removed the work "you'll"