Update source-control.md
the picture does not provide a clear example of the value that should be provided for owner this leaves some ambiguity about the value needed either the full url or just the owner name. 

for reference see this answer on stack overflow:
https://stackoverflow.com/questions/71950079/how-to-connect-azure-synapse-and-an-enterprise-git-repo-using-personal-access-to

I've attempted to leave add this information to make it easier to troubleshoot.