Update enable-vms-monitoring-agent.md
Adding Notes to help others avoid issues when adding Change Tracking. If using the azure portal to add change tracking it will create the DCR rule for you. Also added as notes if you add any data sources to the DCR rule it will break the schema and the inventory and change tracking breaks.