Update automation-manage-send-joblogs-log-analytics.md
deleted ResultType:New entries that were not recorded in JobLogs. Although New is recorded in the Status property of the activity log, it is noted as Created in the ResultType column of the JobLogs table, which contains similar information to the activity log. It is assumed that there was an incorrect mix of information.