Update faq-troubleshooting.md
I am from Microsoft Azure Customer Experience Team and recently I had received a customer case recently where all users lost access to cloud shell with the mentioned error and I could find the process to resolve it as well as an RCA for the same from PG team and internal TSGs. As this is not documented in any MS documents and no proper answers available in Community threads, updating it here.