Update configure-ssl-certificate.md
Added about CAA records in the document at the line 94

++NOTE: Checking CAA Records
Another thing to be noted is CAA record.

CAA is a record in DNS that declares which provider is allowed to issue certificate for this domain. If domain has CAA record value other than GoDaddy, GoDaddy will reject issuing certificate for this domain. We can check CAA in digwebinterface. If CAA record is empty, it will allow all providers like Godaddy, Digicert, etc to issue the certificate. But if CAA record has been set, then it should also be set for GoDaddy for certificate purchase to be allowed.

CAA records for multiple subdomains are checked in sequence. Let's says you are trying to issue a certificate to sub1.my.contoso.com . The initial check will be sub1.my.contoso.com , than my.contoso.com , than contoso.com  until we find a CAA record. If a sublevel contains a value, the root won't be checked and a CAA record at this level should be created.

For standard certificate, CAA record should be - issue "godaddy.com".
For wildcard certificate, CAA record should be - issuewild "godaddy.com".