Update api-management-howto-create-or-invite-developers.md
we need to be very clear as to which specific persona we are talking about in each of the articles.  "developers" is an overloaded term, especially in API eco-system.  you have developers who produce APIs, and producers who consume other people's APIs, and then there are consumers who do not publish anything but are interested in consuming someone else's APIs.
Finally, there is the APIM back-office or admins.
In my commit, i have tried to articulate how a producer can manage its own consumers, since they are the best people to know who they are writing the APIs for.  It could be part of their own org, or outside the org, or outside the company.