Update storage-account-overview.md with workload config table
Update storage-account-overview.md with a workload config table for consumption prior to the full public release of workload presets. Mentions of the portal were removed from this version but will be added again when the feature is available to the general public. Table updated with superscripts and details for superscripts