Add clarity to returned IPs in split-horizon functionality, picture is wrong
In the picture, there is a mistake in the DNS query response. When the Internet client is querying the "VNETA-VM1.contoso.com", it should receive 203.0.113.1. However, in the picture the response shows 56.58.0.1
I added the correct information to the text to increase clarity and accessibility. Please fix the picture as well.