Update max queue storage size
Queue storage documentation says it supports millions of messages up to the capacity of a storage account. Storage account documentation says the capacity of storage accounts is 5 PiB. Yet, this comparison documentation says the max queue storage size is 500 TB.

Trusting the other documentations are accurate, I believe this documentation needs to update the value from 500 TB to 5 PiB. Alternatively, if 500 TB is still the accurate value, more explanation is needed to describe the reasoning.