Add instruction to check that the function Authorization Level is Anonymous
While following the tutorial using C#, I encountered an issue where the function couldn't be called and returned a 401 error when ran in docker. Upon investigating, I discovered that the AuthorizationLevel was set to 'Function', whereas the document assumes it should be set to 'Anonymous' when making the call. I have added an instruction for readers to verify that the authorization level is set to AuthorizationLevel.Anonymous and to update it if necessary