Clean up references to deleted document logic-apps-add-run-inline-code.md
Much of the existing references are from a time when JavaScript was the
only scripting language supported on Logic Apps, so updating the
references to point to the "Execute JavaScript Code" action's
documentation is usually the appropriate move.

Additionally, update some wording to reflect changes in document titles,
section headers, and the surrounding situation with inline code.