Update virtual-network-reference.md
Updating connectivity requirements for TLS chain building. APIM's default certificates are issued by DigiCert, and these endpoints need to be accessible. If they're not accessible, intermediate certs may not get installed correctly.  Removed the text that makes it sound like these are optional.  The APIM default certs are not in customer control, so  these endpoints must be accessible from VNET.