Update tag-support.md
Removed: "> For Azure Front Door Service, you can apply tags when creating the resource, but updating or adding tags is not currently supported. Front Door doesn't support the use of # or : in the tag name."

Tested this in my lab and this appears to be false on all 3 SKUs: Classic, Standard, Premium. I can show to the Contributor in case there is doubt.

Also tested the Traffic Manager / DNS Zone claim and that appears to be correct.