Update howto-configure-fabric-real-time-intelligence.md
For entering the connectionstring, the username/password must be used.

The '$ConnectionString' literal value must be entered in the username field, not the synced secret name. If this is tried, this will lead to a mask error: 'The name must begin and end with a letter or number, and may contain only letters, numbers, hyphens'.

Note that these fields are also entered in https://learn.microsoft.com/en-us/azure/iot-operations/connect-to-cloud/howto-configure-kafka-endpoint?tabs=portal#use-connection-string-for-authentication-to-event-hubs but that text is not as clear as seen here.