Update manage-runtime-environment.md
PowerShell 7.4 is available as an option when creating a new runtime environment but is not noted anywhere in this documentation - so using this PR to add this.

We have been instructed by the PG to advise customers to use a PowerShell 7.4 runtime environment as a suggested solution or workaround in certain Automation Runbook scenarios. So, it would be useful to include PowerShell 7.4 in this documentation.