(AzureCXP) fixes MicrosoftDocs/azure-docs#408947
Updated line no: 118 from "Enable publishing the developer portal for an API Management instance in a VNet by allowing outbound connectivity to blob storage in the West US region. For example, use the Storage.WestUS service tag in an NSG rule. Currently, connectivity to blob storage in the West US region is required to publish the developer portal for any API Management instance." to "Enable publishing the developer portal for an API Management instance in a VNet by allowing outbound connectivity to Azure Storage. For example, use the Storage service tag in an NSG rule. Currently, connectivity to Azure Storage via global or regional service endpoints is required to publish the developer portal for any API Management instance."