(AzureCXP) fixes MicrosoftDocs/azure-docs#415612
Updated line no: 24
From:
A required query parameter wasn't specified for this request.
To:
A required query parameter wasn't specified for this request. Verify that all required parameters are included in your command.

Updated line no: 28
From:
The value provided for one of the HTTP headers wasn't in the correct format.
To:
The value provided for one of the HTTP headers wasn't in the correct format. Check and correct the formatting of your HTTP headers.

Updated line no: 54
From:
The specified resource doesn't exist.
To:
The specified resource doesn't exist. Confirm the resource's URL and ensure it has not been deleted or moved.