Update move-resource-group-and-subscription.md
This stemmed from a case where cx was requesting cost information regarding moving a storage account from one subscription to another.

This documentation is for moving within the same region. There is no publicly available information on what to factor in when considering cost.

When moving resources across regions, there is a documentation on cost. This can be found in Frequently asked questions, "What charges do I incur while using Azure Resource Mover?"
https://azure.microsoft.com/en-us/pricing/details/resource-mover/

https://learn.microsoft.com/en-us/azure/resource-mover/move-region-within-resource-group?toc=%2Fazure%2Fazure-resource-manager%2Fmanagement%2Ftoc.json

When using AzCopy, there is a how to page and a whole page for Estimating the cost of using AzCopy to transfer blobs.

https://learn.microsoft.com/en-us/azure/storage/common/storage-use-azcopy-blobs-copy?toc=%2Fazure%2Fstorage%2Fblobs%2Ftoc.json&bc=%2Fazure%2Fstorage%2Fblobs%2Fbreadcrumb%2Ftoc.json&branch=pr-en-us-259662

https://learn.microsoft.com/en-us/azure/storage/blobs/azcopy-cost-estimation

I've written the question on line 490, "How much does it cost to move Azure resources to a new resource group or subscription within the same region, and what should I consider?" 

I'm not clear on if the answer is same as using Azure Resource Mover.