Update storage-files-identity-ad-ds-overview.md
Adding an example where a group is cloud only and therefore the share level permissions via RBAC will give an error when mounting. If the group contains identities synced from on-premises, then these could be used.

Might be good to change up the wording to make it easier to understand?