(AzureCXP) fixes MicrosoftDocs/azure-docs#421719
Updated line number 303.

From: Although out of scope for this article, for an Azure Data Explorer cluster add the _Reader_ role to the system-assigned managed identity. For more information, at the end of this article select the link _Role assignments for Azure Data Explorer clusters_.

To" From: Although out of scope for this article, if you're querying an Azure Data Explorer cluster, you must cluster add the _Reader_ role to the system-assigned managed identity. For more information, at the end of this article select the link _Role assignments for Azure Data Explorer clusters_.