Alternate Fix for Ingress
There is always a requirement to route the Ingress traffic without using an Azure Firewall and DNAT rule. This solution is not available in any document. Hence I have added the option to help users to use Load Balancer for incoming traffic and also the alternate solution as Application Gateway