Fix formatting of docker compose YAML
Fix a bug where the `emulator` and `azurite` sections under `services`
are missing an indentation level. Without the indentation,
`docker compose` fails with `services must be a mapping`.

Fix a bug where the final `networks` section is missing an indentation
level. Without the indentation, the Markdown parser, and thus the
website, displays both `networks` and its child `eh-emulator` without
indentation. Lack of indentation causes `docker compose` to fail with
`networks must be a mapping`.

Tag the code block as `yaml` to get syntax highlighting.