Update troubleshoot-iot-edge-for-linux-on-windows-common-errors.md
When attempting to connect to the EFLOW VM using the Connect-EflowVm PowerShell cmdlet, the connection fails with an "authentication handshake failed" error.

transport: authentication handshake failed: x509: certificate has expired or is not yet valid

The cause is that the WSSDAgent certificate has expired. The WSSDAgent certificate expires one year after a new installation of EFLOW. Please execute updateWssdCert.ps1 to extend the certificate validity period. Note that updating EFLOW does not extend the validity period of the WSSDAgent certificate.