(AzureCXP) fixes MicrosoftDocs/azure-docs#438834
Updated document by adding line no: 29-31 with "What is the difference between a datacenter and an availability zone?
An availability zone is a logical grouping of one or more physically separate datacenters within a region. Each availability zone is built in a way that if something goes wrong in one (like a power outage or network issue), the others keep working. A single datacenter doesn’t offer this level of protection on its own."