Update service-limits.md for ACS Chat to introduce new limits
ACS Chat is going to add a new limit to throttle the usage of create chat thread and add chat participant APIs based on the requests from the same resource Id. Although the throttle limit exists on number of chat thread created per user but since a resource has the ability to create many such users, so we need to put a throttle limit at resource Id level.