Update redundancy-migration.md
There were a couple typos with the cli commands given and also, I guess some flags there have one - and some have two. It may be easier to show the two because then the flag is resource-group instead of just g. Just a suggestion there but the spaces before -'s need to be fixed because the command won't work like that.