Add redirection that was mistakenly added to virtual-desktop
This redirection entry was added in June 2023 to the redirection file for Azure Virtual Desktop. That content and its redirection file are now in windows-cloud-pr, so I'm restoring this redirection for disconnected-mod.md to azure-docs-pr. I didn't see a definite match for /private-5g-core/ in ./redirects/, so I'm choosing the default redirection file.