Update public-ip-address-prefix.md
I added a line to the VM, but in a couple of cases we have seen customers demand that any access to their network has to be with public IP space.  To facilitate this in Azure we create the Public IP prefix, add it to the VNET, associate it with the Subnet, Assign it to the VM.  BGP Routing will then send the prefix to the gateway and allow internal routing of the range.