Update note on subnet delegation requirements
Because the Microsoft.Web/serverFarms resource provider shows on the portal even if the service is not registered for the subscription, it might become difficult to troubleshoot. The error message when it is not registered says the setup was not able to set the delegation, instead of saying the resource is not registered. Most of my customers assume it is a permission issue, not a service registration issue.