Update overview-inbound-outbound-ips.md
Change "Get a static inbound IP" to -> "Get a dedicated inbound IP" 

based on Supportability ask for clarification, 
that the difference is this feature sets a *dedicated* address,
by default the inbound IP is already static. 

So the previous verbiage is confusing, and my proposed change is more accurate. 

Validated with product SMEs as good change.